加入了派单算法管理控制

+++ sosotaxi/dispatch-service/src/main/java/com/apicaller/sosotaxi/service/impl/DispatchServiceImpl.java
@@ -15,6 +15,65 @@ import java.util.Set;
 @Component
 public class DispatchServiceImpl implements DispatchService {
 
+    /**
+     * 派单模式
+     */
+    public enum dispatchMethod{
+        /**
+         * 抢单模式
+         */
+        GROUPDISPATCH,
+        /**
+         * 自动安排模式
+         */
+        AUTOARRANGE
+    }
+
+    /**
+     * 当前策略
+     */
+    private dispatchMethod currentMethod = dispatchMethod.GROUPDISPATCH;
+
+    /**
+     * 更改策略
+     * @param method
+     * @return
+     */
+    public Boolean changeDispatchMethod(String method){
+        if(method.equals(dispatchMethod.AUTOARRANGE.toString())){
+            currentMethod = dispatchMethod.AUTOARRANGE;
+        }
+        else if(method.equals(dispatchMethod.GROUPDISPATCH.toString())){
+            currentMethod = dispatchMethod.GROUPDISPATCH;
+        }
+        else{
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * 抢单模式下一单的最大分配司机数
+     */
+    static private int arrangeCount = 8;
+
+    /**
+     * 设置最大分配司机数
+     * 仅当模式为抢单模式时有效
+     * @param arrangeCount
+     */
+    public void setArrangeCount(int arrangeCount){
+        DispatchServiceImpl.arrangeCount = arrangeCount;
+    }
+
+    /**
+     * 设置最大分配司机数
+     * @return
+     */
+    public int getArrangeCount(){
+        return DispatchServiceImpl.arrangeCount;
+    }
+
     @Resource
     private InfoCacheServiceImpl infoCacheService;
 
@@ -25,6 +84,10 @@ public class DispatchServiceImpl implements DispatchService {
     @Override
     @Scheduled(fixedDelay = 5000)
     public void dispatch() throws Exception {
+        //仅当当前策略为抢单模式时才执行定时处理方法
+        if(currentMethod!=dispatchMethod.GROUPDISPATCH){
+            return;
+        }
         List<UnsettledOrder> orders = infoCacheService.getAllUOrder();
         Set<String> driverIds = infoCacheService.getAllDrivers();
         infoCacheService.deleteDurationSets(infoCacheService.getAllDurationSetKeys());
@@ -52,7 +115,7 @@ public class DispatchServiceImpl implements DispatchService {
             }
             Set<String> driversToDispatch = null;
             try {
-                driversToDispatch = infoCacheService.getDriverIdByRank(order.getOrderId());
+                driversToDispatch = infoCacheService.getDriverIdByRank(order.getOrderId(), arrangeCount);
             }catch (Exception e){
                 e.printStackTrace();
             }
